requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref
asn1crypto
Brotli
certifi
cffi
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
cssselect2==0.7.0
distro
Django==5.1.6
django-allauth==65.4.1
h11
html5lib
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
openai==1.65.1
oscrypto==1.3.0
pillow
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.8.0
pypdf==5.3.0
PySocks
python-bidi==0.6.6
PyYAML
qrcode==8.0
reportlab==4.3.1
requests
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tailwindcss==0.0.1
tinycss2
tqdm
typing_extensions
tzlocal==5.3
uritools==4.0.3
urllib3
webencodings==0.5.1
xhtml2pdf==0.2.17

dotenv~=0.9.9
python-dotenv~=1.0.1